requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
APScheduler==3.11.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
frozenlist==1.7.0
idna==3.10
iniconfig==2.1.0
jalali_core==1.0.0
jdatetime==5.2.0
multidict==6.6.4
mysql-connector-python==9.4.0
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
Pygments==2.19.2
pyTelegramBotAPI==4.28.0
pytest==8.4.1
requests==2.32.5
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wheel==0.45.1
yarl==1.20.1
